Refactor plotting functions to plots module
